chore(atomic-bomb): Compile atomic bomb so we can build custom components
Unfortunately, it won't be possible for us to just add class names to semantic's components to style them using Atomic Bomb. That's because part of the styling happens to internal elements of these components, which we can't set classNames for.

We don't want to have to copy/paste all design specs from Atomic Bomb to Orion, so we'll recompile Atomic Bomb here, which will allow reusing the existing utilities to build custom components when we necessary.