fix(button): Normalize Button's icon prop
Semantic's icon prop frequently throws warnings for us because it expects icon names from a certain icon library, but we use another. The way out has been to pass the icon as a class name instead of just the icon's name, but that's annoying.

Since we're overriding these components now, it's our chance to automatically do this tansformation from icon name to `{ className: iconName }`, hiding this from Orion's users.