fix(checkbox): Fix spacing between checkboxes
I've added spacing between the checkboxes, which should be 16px. But I forgot that the browser automatically merges margins, causing the spacing to become 14px, due to the 2px margin from each checkbox.

I've fixed this by setting the spacing to 20px margin instead, which becomes 16px when the preexisting margins are not considered. Now it actually looks like the prototype at Invision.

Unfortunately we don't have the 20 as a valid spacing in atomic bomb, so I've added the rule directly.