feature(datepicker): Remove horizontal spacing from datepickers
This is not as easy to do as it normally is, since this spacing is automatically added via javascript by react-dates. The only way to remove it, while still keeping the spacing between months (in pickers with multiple months), was to add a wrapper with negative margin to both datepickers.