fix(infoIcon): Pass props over to the icon instead of the popup
The main element here is the icon, not the popup, and it's the one that may need to be customized (with margin, for example). If we find a need to customize the popup, we can add an extra prop for that. For now people will just use `content` for the popup's content and the rest for the icon itself.